+++ src/index.js
@@ -1,5 +1,5 @@
 import ReactDOM from 'react-dom';
-import { App } from './components/App';
+import { App } from './components/App/App';
 
 import styles from './index.scss';
