+++ src/views/Home/Home.js
@@ -1,10 +1,6 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React from "react";
+import PropTypes from "prop-types";
 
-export const Home = () => (
-  <div>
-    Home page
-  </div>
-);
-
-Home.prop
+export const Home = () => {
+  return <div>Home page.</div>;
+};
